+++ jest.setup.js
@@ -1,7 +1,7 @@
+/* eslint-disable no-unused-expressions */
 import mockDevices from './src/mockDevices';
 
 // Mock Web Media APIs
-
 global.MediaStream = jest.fn(function MockMediaStream(tracksOrStream = []) {
   let tracks;
 
@@ -32,11 +32,9 @@ global.MediaStream = jest.fn(function MockMediaStream(tracksOrStream = []) {
   });
 });
 
-global.navigator = {
-  mediaDevices: {
-    enumerateDevices: () => Promise.resolve(mockDevices),
-  },
-  userAgent: ''
+global.navigator.mediaDevices = {
+  enumerateDevices: () => Promise.resolve(mockDevices),
+  getDisplayMedia: () => Promise.resolve(new MediaStream()),
 };
 
 expect.extend({
@@ -68,14 +66,12 @@ Received: ${this.utils.printReceived(received && received.getTracks())}`;
   },
 });
 
-global.document = {
-  createElement: (type) => {
-    const elem = {};
+global.document.createElement = (type) => {
+  const elem = {};
 
-    if (type === 'audio' || type === 'video') {
-      elem.setSinkId = () => {};
-    }
+  if (type === 'audio' || type === 'video') {
+    elem.setSinkId = () => {};
+  }
 
-    return elem;
-  },
+  return elem;
 };