+++ src/ActivitiesSDKAdapter.js
@@ -2,17 +2,19 @@ import {constructHydraId, deconstructHydraId} from '@webex/common';
 import {ActivitiesAdapter} from '@webex/component-adapter-interfaces';
 import {
   from,
+  defer,
   Observable,
   ReplaySubject,
-  defer,
 } from 'rxjs';
 import {
   catchError,
+  concatMap,
   map,
   tap,
 } from 'rxjs/operators';
 
 import logger from './logger';
+import cache from './cache';
 
 /**
  * An activity a person performs in Webex.
@@ -22,28 +24,39 @@ import logger from './logger';
  */
 
 /**
- * Extracts JSON card and files from server activity.
+ * Extracts JSON cards from server activity.
  *
  * @private
  * @param {object} sdkActivity  SDK activity object
- * @returns {Array} Array of attachments.
+ * @returns {Array} Array of cards.
  */
-function parseSDKAttachments(sdkActivity) {
-  const {attachments = []} = sdkActivity;
+function parseSDKCards(sdkActivity) {
+  let cards = [];
 
-  try {
-    if (sdkActivity.object && sdkActivity.object.cards) {
-      sdkActivity.object.cards.forEach((c) => {
-        const card = JSON.parse(c);
+  if (sdkActivity.object && sdkActivity.object.cards) {
+    cards = sdkActivity.object.cards.map((c) => {
+      let card;
 
-        attachments.push({contentType: 'application/vnd.microsoft.card.adaptive', content: card});
-      });
-    }
-  } catch (err) {
-    logger.warn('ACTIVITY', undefined, 'parseSDKAttachments()', `Unable parse attachments for activity with id "${sdkActivity.id}"`, err);
+      try {
+        card = JSON.parse(c);
+      } catch (err) {
+        logger.warn('ACTIVITY', sdkActivity.id, 'parseSDKCards()', ['Unable parse card', c], err);
+
+        card = {
+          type: 'AdaptiveCard',
+          version: '1.0',
+          body: [{
+            type: 'TextBlock',
+            text: 'This card could not be parsed.',
+          }],
+        };
+      }
+
+      return card;
+    });
   }
 
-  return attachments;
+  return cards;
 }
 
 /**
@@ -53,14 +66,16 @@ function parseSDKAttachments(sdkActivity) {
  * @param {object} sdkActivity  SDK activity object
  * @returns {Activity} Adapter activity object
  */
-function fromSDKActivity(sdkActivity) {
+export function fromSDKActivity(sdkActivity) {
   return {
-    ID: sdkActivity.id ? constructHydraId('message', sdkActivity.id) : sdkActivity.ID,
-    roomID: sdkActivity.target ? constructHydraId('room', sdkActivity.target.id) : sdkActivity.roomID,
-    personID: sdkActivity.actor ? constructHydraId('person', sdkActivity.actor.id) : sdkActivity.personID,
+    displayHeader: true,
+    ID: sdkActivity.id ? constructHydraId('MESSAGE', sdkActivity.id) : sdkActivity.ID,
+    roomID: sdkActivity.target ? constructHydraId('ROOM', sdkActivity.target.id) : sdkActivity.roomID,
+    personID: sdkActivity.actor ? constructHydraId('PEOPLE', sdkActivity.actor.id) : sdkActivity.personID,
     text: sdkActivity.object ?
       (sdkActivity.object.content || sdkActivity.object.displayName) : sdkActivity.text,
-    attachments: parseSDKAttachments(sdkActivity),
+    cards: parseSDKCards(sdkActivity),
+    attachments: sdkActivity.attachments || [],
     created: sdkActivity.published ? sdkActivity.published : sdkActivity.created,
   };
 }
@@ -89,12 +104,37 @@ export default class ActivitiesSDKAdapter extends ActivitiesAdapter {
    * @private
    */
   async fetchActivity(activityID) {
-    logger.debug('ACTIVITY', activityID, 'fetchActivity()', ['called with', {activityID}]);
+    logger.debug('ACTIVITY', 'fetchActivity()', activityID);
 
     const {id} = deconstructHydraId(activityID);
     const service = 'conversation';
-    const resource = `activities/${id}`;
+    const resource = `activities/${encodeURIComponent(id)}`;
+
+    if (cache.has(id)) {
+      return cache.get(id);
+    }
+
     const {body} = await this.datasource.request({service, resource});
+
+    cache.set(id, body);
+
+    return body;
+  }
+
+  /**
+   * Fetches a conversation from the API
+   *
+   * @param {string} conversationID  Id of the conversation for which to fetch data
+   * @returns {Promise} Information about the conversation of the given ID
+   *
+   * @private
+   */
+  async fetchConversation(conversationID) {
+    const {id} = deconstructHydraId(conversationID);
+    const method = 'GET';
+    const api = 'conversation';
+    const resource = `conversations/${encodeURIComponent(id)}`;
+    const {body} = await this.datasource.request({method, api, resource});
 
     return body;
   }
@@ -106,11 +146,11 @@ export default class ActivitiesSDKAdapter extends ActivitiesAdapter {
    * @returns {external:Observable.<Activity>} Observable stream that emits activity data
    */
   getActivity(ID) {
-    logger.debug('ACTIVITY', ID, 'getActivity()', ['called with', {ID}]);
+    logger.debug('ACTIVITY', 'getActivity()', ID, []);
 
     if (!(ID in this.activityObservables)) {
       // use ReplaySubject cause we don't need to set an initial value
-      this.activityObservables[ID] = new ReplaySubject(1);
+      this.activityObservables[ID] = new ReplaySubject();
 
       defer(() => this.fetchActivity(ID)).pipe(
         map(fromSDKActivity),
@@ -137,105 +177,101 @@ export default class ActivitiesSDKAdapter extends ActivitiesAdapter {
    * @returns {Observable.<object>} Observable stream that emits data of the newly created action
    */
   postAction(activityID, inputs) {
-    logger.debug('ATTACHMENT-ACTION', undefined, 'postAction()', ['called with', {activityID, inputs}]);
+    logger.debug('ACTION', undefined, 'postAction()', ['called with', {activityID, inputs}]);
 
-    const action$ = from(this.datasource.attachmentActions.create({
-      type: 'submit',
-      messageId: activityID,
-      inputs,
-    })).pipe(
-      map((action) => ({
-        actionID: action.id,
-        activityID: action.messageId,
-        inputs: action.inputs,
-        roomID: action.roomId,
-        personID: action.personId,
-        type: action.type,
-        created: action.created,
-      })),
-      tap((action) => {
-        logger.debug('ATTACHMENT-ACTION', action.actionID, 'postAction()', ['emitting posted attachment action', action]);
+    return from(this.fetchActivity(activityID)).pipe(
+      concatMap(async (parentActivity) => {
+        const encryptedInputs = await this.datasource.internal.encryption
+          .encryptText(parentActivity.encryptionKeyUrl, JSON.stringify(inputs));
+
+        return this.datasource.internal.conversation.cardAction(
+          parentActivity.target,
+          {inputs: encryptedInputs},
+          parentActivity,
+        );
       }),
+      tap((action) => {
+        logger.debug('ACTION', action.id, 'postAction()', ['emitting posted action', action]);
+      }),
+      map(fromSDKActivity),
       catchError((err) => {
-        logger.error('ATTACHMENT-ACTION', undefined, 'postAction()', `Unable to create an attachment for activity with id "${activityID}"`, err);
+        logger.error('ACTION', undefined, 'postAction()', `Unable to create an action for activity with id "${activityID}"`, err);
         throw err;
       }),
     );
+  }
 
-    return action$;
+  /**
+   * Encrypts the cards of an activity and returns a promise to the array of encrypted cards
+   *
+   * @private
+   * @param {Activity} activity  The activity that contains an array of adaptive cards
+   * @returns {Promise.<Array.<string>>} Promise that resolves to the array of encrypted cards or rejects if cards cannot be encrypted
+   */
+  async encryptCards(activity) {
+    logger.debug('ACTIVITY', activity.ID, 'encryptCards()', ['called with', {activity}]);
+
+    const conversation = await this.fetchConversation(activity.roomID);
+
+    return Promise.all(activity.cards.map((card) => (
+      this.datasource.internal.encryption.encryptText(
+        conversation.encryptionKeyUrl,
+        JSON.stringify(card),
+      )
+    ))).catch((err) => {
+      logger.error('ACTIVITY', activity.ID, 'encryptCards()', 'Unable to encrypt card', err);
+      throw err;
+    });
   }
 
   /**
    * Posts an activity and returns an observable to the new activity data
    *
-   * @param {object} activity  The activity to post
+   * @param {Activity} activity  The activity to post
    * @returns {Observable.<Activity>} Observable that emits the posted activity (including id)
    */
   postActivity(activity) {
     logger.debug('ACTIVITY', undefined, 'postActivity()', ['called with', {activity}]);
-    const card = this.getAdaptiveCard(activity);
 
-    const object = card && {
-      cards: [JSON.stringify(card)],
-      displayName: activity.text,
+    const doPost = async () => {
+      const {id, cluster = 'us'} = deconstructHydraId(activity.roomID);
+      const hasCards = this.hasAdaptiveCards(activity);
+      const object = hasCards
+        ? ({cards: await this.encryptCards(activity), displayName: activity.text})
+        : activity.text;
+
+      return this.datasource.internal.conversation.post({id, cluster}, object);
     };
 
-    const {id, cluster = 'us'} = deconstructHydraId(activity.roomID);
-
-    const activity$ = from(this.datasource.internal.conversation.post(
-      {
-        id,
-        cluster,
-      },
-      object || activity.text,
-    )).pipe(
+    return defer(doPost).pipe(
       map(fromSDKActivity),
       catchError((err) => {
         logger.error('ACTIVITY', undefined, 'postActivity()', ['Unable to post activity', activity], err);
         throw err;
       }),
     );
-
-    return activity$;
   }
 
   /**
-   * A function that checks whether or not an Activity object contains a card attachment.
+   * A function that checks whether or not an Activity object contains at least one adaptive card.
    *
    * @param {Activity} activity  Activity object
-   * @returns {boolean} True if received Activity object contains a card attachment
+   * @returns {boolean} True if received Activity object contains at least one adaptive card
    */
   // eslint-disable-next-line class-methods-use-this
-  hasAdaptiveCard(activity) {
-    return !!(activity.attachments && activity.attachments[0] && activity.attachments[0].contentType === 'application/vnd.microsoft.card.adaptive');
+  hasAdaptiveCards(activity) {
+    return activity.cards.length > 0;
   }
 
   /**
    * A function that returns adaptive card data of an Activity object.
    *
    * @param {Activity} activity  Activity object
+   * @param {number} cardIndex  Index of the card to get
    * @returns {object|undefined} Adaptive card data object
    */
   // eslint-disable-next-line class-methods-use-this
-  getAdaptiveCard(activity) {
-    const hasCard = this.hasAdaptiveCard(activity);
-
-    return hasCard ? activity.attachments[0].content : undefined;
-  }
-
-  /**
-   * A function that attaches an adaptive card to an Activity object.
-   *
-   * @param {Activity} activity  The activity to post
-   * @param {object} card  The card attachment
-   */
-  // eslint-disable-next-line class-methods-use-this
-  attachAdaptiveCard(activity, card) {
-    const mutableActivity = activity;
-
-    mutableActivity.attachments = [{
-      contentType: 'application/vnd.microsoft.card.adaptive',
-      content: card,
-    }];
+  getAdaptiveCard(activity, cardIndex) {
+    return activity.cards[cardIndex];
   }
 }