+++ src/ActivitiesSDKAdapter.test.js
@@ -3,18 +3,38 @@ import {last} from 'rxjs/operators';
 
 import ActivitiesSDKAdapter from './ActivitiesSDKAdapter';
 import createMockSDK, {
-  mockSDKCardActivity, created, serverActivity, activityID, roomID, personID, ID, actorID, targetID,
+  created,
+  sdkActivity,
+  sdkConversation,
+  activityID,
+  roomID,
+  personID,
+  ID,
+  actorID,
+  targetID,
 } from './mockSdk';
 
 describe('Activities SDK Adapter', () => {
   let mockSDK;
   let activitiesSDKAdapter;
 
-  const activityWithoutCard = {
+  const activityWithoutCards = {
     ID,
     roomID,
     text: 'text1',
     personID,
+    cards: [],
+    attachments: [],
+    created,
+  };
+
+  const activityWithCards = {
+    ID,
+    roomID,
+    text: 'text1',
+    personID,
+    cards: [JSON.parse(sdkActivity.object.cards[0])],
+    attachments: [],
     created,
   };
 
@@ -31,15 +51,15 @@ describe('Activities SDK Adapter', () => {
   describe('getActivity()', () => {
     beforeEach(() => {
       activitiesSDKAdapter.fetchActivity = jest.fn(
-        () => Promise.resolve(serverActivity),
+        () => Promise.resolve(sdkActivity),
       );
     });
 
-    test('returns an observable', () => {
+    it('returns an observable', () => {
       expect(isObservable(activitiesSDKAdapter.getActivity(activityID))).toBeTruthy();
     });
 
-    test('emits activity details on subscription', (done) => {
+    it('emits activity details on subscription', (done) => {
       activitiesSDKAdapter.getActivity(activityID).subscribe(
         (activity) => {
           expect(activity).toMatchObject({
@@ -47,30 +67,26 @@ describe('Activities SDK Adapter', () => {
             roomID,
             text: 'text',
             personID,
-            attachments: [
-              {
-                contentType: 'application/vnd.microsoft.card.adaptive',
-                content: {
-                  $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
-                  type: 'AdaptiveCard',
-                  version: '1.2',
-                  body: [
-                    {
-                      type: 'TextBlock',
-                      text: 'Adaptive Cards',
-                      size: 'large',
-                    },
-                  ],
-                  actions: [
-                    {
-                      type: 'Action.OpenUrl',
-                      url: 'http://adaptivecards.io',
-                      title: 'Learn More',
-                    },
-                  ],
+            cards: [{
+              $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
+              type: 'AdaptiveCard',
+              version: '1.2',
+              body: [
+                {
+                  type: 'TextBlock',
+                  text: 'Adaptive Cards',
+                  size: 'large',
                 },
-              },
-            ],
+              ],
+              actions: [
+                {
+                  type: 'Action.OpenUrl',
+                  url: 'http://adaptivecards.io',
+                  title: 'Learn More',
+                },
+              ],
+            }],
+            attachments: [],
             created,
           });
           done();
@@ -78,7 +94,7 @@ describe('Activities SDK Adapter', () => {
       );
     });
 
-    test('throws an error on invalid activity ID', (done) => {
+    it('throws an error on invalid activity ID', (done) => {
       const sdkError = new Error('Could not find activity with ID "badActivityID"');
 
       activitiesSDKAdapter.fetchActivity = jest.fn(() => Promise.reject(sdkError));
@@ -94,11 +110,37 @@ describe('Activities SDK Adapter', () => {
   });
 
   describe('postActivity()', () => {
-    test('emits the posted Activity object', (done) => {
+    beforeEach(() => {
+      activitiesSDKAdapter.fetchConversation = jest.fn(
+        () => Promise.resolve(sdkConversation),
+      );
+    });
+
+    it('emits the posted Activity object', (done) => {
       const activityData = {
         roomID,
         personID,
         text: 'text',
+        cards: [{
+          $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
+          type: 'AdaptiveCard',
+          version: '1.2',
+          body: [
+            {
+              type: 'TextBlock',
+              text: 'Adaptive Cards',
+              size: 'large',
+            },
+          ],
+          actions: [
+            {
+              type: 'Action.OpenUrl',
+              url: 'http://adaptivecards.io',
+              title: 'Learn More',
+            },
+          ],
+        }],
+        attachments: [],
       };
 
       activitiesSDKAdapter.datasource.internal.conversation.post = jest.fn(
@@ -109,6 +151,7 @@ describe('Activities SDK Adapter', () => {
           },
           object: {
             displayName: 'text',
+            cards: [JSON.stringify(activityData.cards[0])],
           },
           target: {
             id: targetID,
@@ -124,153 +167,7 @@ describe('Activities SDK Adapter', () => {
           roomID,
           personID,
           created,
-        });
-        done();
-      });
-    });
-
-    test('emits the sdk error when messages.create returns a rejected promise', (done) => {
-      const sdkError = new Error('sdk-error');
-
-      activitiesSDKAdapter.datasource.internal.conversation.post = jest.fn(
-        () => Promise.reject(sdkError),
-      );
-
-      activitiesSDKAdapter.postActivity({roomID: ''}).subscribe(
-        () => {
-          done.fail('Posted an activity instead of returning error');
-        },
-        (error) => {
-          expect(error).toBe(sdkError);
-          done();
-        },
-      );
-    });
-  });
-
-  describe('postAction()', () => {
-    test('emits the posted action object', (done) => {
-      const inputs = {
-        firstName: 'My first name',
-        lastname: 'My last name',
-      };
-
-      activitiesSDKAdapter.postAction(activityID, inputs).pipe(last()).subscribe((action) => {
-        expect(action).toMatchObject({
-          actionID: 'actionID',
-          activityID: 'activityID',
-          inputs: {
-            firstName: 'My first name',
-            lastName: 'My last name',
-          },
-          roomID: 'roomID',
-          personID: 'personID',
-          type: 'submit',
-          created,
-        });
-        done();
-      });
-    });
-
-    test('emits the sdk error when attachmentActions.create returns a rejected promise', (done) => {
-      const sdkError = new Error('sdk-error');
-
-      activitiesSDKAdapter.datasource.attachmentActions.create = jest.fn(
-        () => Promise.reject(sdkError),
-      );
-
-      activitiesSDKAdapter.postAction({}).subscribe(
-        () => {
-          done.fail('Created attachment action instead of returning error');
-        },
-        (error) => {
-          expect(error).toBe(sdkError);
-          done();
-        },
-      );
-    });
-  });
-
-  describe('hasAdaptiveCard()', () => {
-    test('returns true if activity object has a card attachment', () => {
-      const hasCard = activitiesSDKAdapter.hasAdaptiveCard(mockSDKCardActivity);
-
-      expect(hasCard).toBeTruthy();
-    });
-
-    test('returns false if activity object does not have a card attachment', () => {
-      const hasCard = activitiesSDKAdapter.hasAdaptiveCard(activityWithoutCard);
-
-      expect(hasCard).toBeFalsy();
-    });
-  });
-
-  describe('getAdaptiveCard()', () => {
-    test('returns the card object if the activity object has a card attachment', () => {
-      const cardAttachment = activitiesSDKAdapter.getAdaptiveCard(mockSDKCardActivity);
-
-      expect(cardAttachment).toMatchObject({
-        $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
-        type: 'AdaptiveCard',
-        version: '1.2',
-        body: [
-          {
-            type: 'TextBlock',
-            text: 'Adaptive Cards',
-            size: 'large',
-          },
-        ],
-        actions: [
-          {
-            type: 'Action.OpenUrl',
-            url: 'http://adaptivecards.io',
-            title: 'Learn More',
-          },
-        ],
-      });
-    });
-
-    test('returns undefined if Activity object has a card attachment', () => {
-      const cardAttachment = activitiesSDKAdapter.getAdaptiveCard(activityWithoutCard);
-
-      expect(cardAttachment).toBeUndefined();
-    });
-  });
-
-  describe('attachAdaptiveCard()', () => {
-    test('add an adaptive card attachment to the activity', () => {
-      const activity = {
-        roomID: 'roomID3',
-        text: 'text3',
-      };
-      const cardToAttach = {
-        $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
-        type: 'AdaptiveCard',
-        version: '1.2',
-        body: [
-          {
-            type: 'TextBlock',
-            text: 'Adaptive Cards',
-            size: 'large',
-          },
-        ],
-        actions: [
-          {
-            type: 'Action.OpenUrl',
-            url: 'http://adaptivecards.io',
-            title: 'Learn More',
-          },
-        ],
-      };
-
-      activitiesSDKAdapter.attachAdaptiveCard(activity, cardToAttach);
-
-      expect(activity).toMatchObject({
-        roomID: 'roomID3',
-        text: 'text3',
-        attachments: [{
-          contentType: 'application/vnd.microsoft.card.adaptive',
-          content: {
+          cards: [{
             $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
             type: 'AdaptiveCard',
             version: '1.2',
@@ -288,9 +185,134 @@ describe('Activities SDK Adapter', () => {
                 title: 'Learn More',
               },
             ],
-          },
-        }],
+          }],
+          attachments: [],
+        });
+        done();
       });
+    });
+
+    it('emits the sdk error when messages.create returns a rejected promise', (done) => {
+      const sdkError = new Error('sdk-error');
+
+      activitiesSDKAdapter.datasource.internal.conversation.post = jest.fn(
+        () => Promise.reject(sdkError),
+      );
+
+      activitiesSDKAdapter.postActivity({roomID: '', cards: [], attachments: []}).subscribe(
+        () => {
+          done.fail('Posted an activity instead of returning error');
+        },
+        (error) => {
+          expect(error).toBe(sdkError);
+          done();
+        },
+      );
+    });
+  });
+
+  describe('postAction()', () => {
+    beforeEach(() => {
+      activitiesSDKAdapter.fetchActivity = jest.fn(
+        () => Promise.resolve(sdkActivity),
+      );
+    });
+    it('emits the posted action object', (done) => {
+      const inputs = {
+        firstName: 'My first name',
+        lastname: 'My last name',
+      };
+
+      activitiesSDKAdapter.datasource.internal.conversation.cardAction = jest.fn(
+        () => Promise.resolve({
+          ID,
+          actor: {
+            id: actorID,
+          },
+          object: {
+            displayName: 'text',
+          },
+          target: {
+            id: targetID,
+          },
+          published: created,
+        }),
+      );
+
+      activitiesSDKAdapter.postAction(activityID, inputs).subscribe((action) => {
+        expect(action).toMatchObject({
+          ID,
+          text: 'text',
+          roomID,
+          personID,
+          created,
+        });
+        done();
+      });
+    });
+
+    it('emits the sdk error when internal.conversation.cardAction returns a rejected promise', (done) => {
+      const sdkError = new Error('sdk-error');
+
+      activitiesSDKAdapter.datasource.internal.conversation.cardAction = jest.fn(
+        () => Promise.reject(sdkError),
+      );
+
+      activitiesSDKAdapter.postAction(activityID, {x: 1, y: 2}).subscribe(
+        () => {
+          done.fail('Created card action instead of returning error');
+        },
+        (error) => {
+          expect(error).toBe(sdkError);
+          done();
+        },
+      );
+    });
+  });
+
+  describe('hasAdaptiveCards()', () => {
+    it('returns true if activity object has at least one adaptive card', () => {
+      const hasCards = activitiesSDKAdapter.hasAdaptiveCards(activityWithCards);
+
+      expect(hasCards).toBeTruthy();
+    });
+
+    it('returns false if activity object does not have at least one adaptive card', () => {
+      const hasCard = activitiesSDKAdapter.hasAdaptiveCards(activityWithoutCards);
+
+      expect(hasCard).toBeFalsy();
+    });
+  });
+
+  describe('getAdaptiveCard()', () => {
+    it('returns the first card object if the activity object has at least one adaptive card', () => {
+      const card = activitiesSDKAdapter.getAdaptiveCard(activityWithCards, 0);
+
+      expect(card).toMatchObject({
+        $schema: 'http://adaptivecards.io/schemas/adaptive-card.json',
+        type: 'AdaptiveCard',
+        version: '1.2',
+        body: [
+          {
+            type: 'TextBlock',
+            text: 'Adaptive Cards',
+            size: 'large',
+          },
+        ],
+        actions: [
+          {
+            type: 'Action.OpenUrl',
+            url: 'http://adaptivecards.io',
+            title: 'Learn More',
+          },
+        ],
+      });
+    });
+
+    it('returns undefined if Activity object does not have at least one card', () => {
+      const card = activitiesSDKAdapter.getAdaptiveCard(activityWithoutCards, 0);
+
+      expect(card).toBeUndefined();
     });
   });
 });